% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-traces.R
\name{clear_traces}
\alias{clear_traces}
\title{Clear previous traces}
\usage{
clear_traces()
}
\description{
Traces are by default appended to previous traces. This function can be used
to clean those previous ones, to enable subsequent calls to generate new
traces that are not appended to previous ones.
}
