% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject-code.R
\name{inject_code}
\alias{inject_code}
\title{Injects code into an existing function.}
\usage{
inject_code(fun, code, where = c("onentry", "onexit", "onerror"))
}
\arguments{
\item{fun}{a function into which the code should be inserted.}

\item{code}{an expression that should be inserted.}

\item{where}{a indicates at which point of the \code{fun} invocation the given
\code{code} shall run.}
}
\description{
The \code{code} argument is recorded without evaluation and it will be inserted
at the beginning of the \code{fun}. If \code{where} is not \code{onentry} it will be
wrapped in \code{on.exit} call. If \code{where} is \code{onerror}, the wrapped code will
further make the \code{code} to run only in the case the function failed with an
error condition.
}
