% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject-tracer.R
\name{inject_tracer}
\alias{inject_tracer}
\title{Inject parameter tracer into one function}
\usage{
inject_tracer(f, e)
}
\arguments{
\item{f}{Name of function as character string}

\item{e}{Environment in which function is defined; generally
\code{as.environment("package:<pkg_name>")}.}
}
\value{
Nothing (will error on fail).
}
\description{
Inject parameter tracer into one function
}
