% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-traces.R
\name{load_traces}
\alias{load_traces}
\title{Load traces of parameter types}
\usage{
load_traces(quiet = FALSE)
}
\arguments{
\item{quiet}{If \code{FALSE}, issue message when no traces found.}
}
\description{
Load traces of parameter types
}
