% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace-package.R
\name{trace_package}
\alias{trace_package}
\title{Trace all parameters for all functions in a specified package}
\usage{
trace_package(package = NULL)
}
\arguments{
\item{Package}{Name of package to be traced (as character value)}
}
\value{
A \code{data.frame} of data on every parameter of every function as
specified in code provided in package examples.
}
\description{
Trace all parameters for all functions in a specified package
}
