% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-and-clear-traces.R
\name{load_traces}
\alias{load_traces}
\title{Load traces of parameter types}
\usage{
load_traces(files = FALSE, quiet = FALSE)
}
\arguments{
\item{files}{If \code{TRUE}, return paths to all temporary files holding trace
data.}

\item{quiet}{If \code{FALSE}, issue message when no traces found.}
}
\value{
A 'data.frame' of traces, including names of functions and
parameters, and values of each parameter traced in both unevaluated and
evaluated forms.
}
\description{
Load traces of parameter types
}
\examples{
f <- function (x, y, z, ...) {
    x * x + y * y
}
inject_tracer (f)
val <- f (1:2, 3:4 + 0., a = "blah")
x <- load_traces ()
print (x)

# Traces should always be "uninjected":
uninject_tracer (f)
# Traces may also be removed:
clear_traces ()
}
