% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject-tracer.R
\name{uninject_tracer}
\alias{uninject_tracer}
\title{Remove parameter tracer from one function}
\usage{
uninject_tracer(f)
}
\arguments{
\item{f}{A function (that is, an object of class "function", and not a
character string).}
}
\value{
Logical value indicating whether or not tracer was able to be removed
("uninjected").
}
\description{
This function removes traces previous injected into functions with the
\link{inject_tracer} function.
}
